+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## [1.0.0+69]
 
-- Add gallery and story management features to UserAccounts model; update version to 1.0.0+70.
+- feat: Enhance NotificationsController with category parameters and update version to 1.0.0+74